pyproject.toml
[tool.poetry]
name = "notion_bot"
version = "0.1.0"
description = ""
authors = ["Vlad-Vargan <blach.smith.2000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.10"
python-dotenv = "^0.19.2"
loguru = "^0.5.3"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
black = {version = "^21.12b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
